11 to 15 problems compleated.

+++ 01-Intro/problem-12.cpp
@@ -1,8 +1,10 @@
+//Program that will define a constant using “CONST” and print the value.
+
 #include<bits/stdc++.h>
 using namespace std;
 int main()
 {
-    const int pi = 3.1416;
+    const float pi = 3.1416;
     cout<<"The value of PI : "<<pi;
     return 0;
 }
